refactor: pages and layouts, components moved to proper places
