+++ src/app/profile/[username]/page.tsx
@@ -1,51 +1,299 @@
+"use client"
+
 import {
   getProfileByUsername,
-  getUserLikedPosts,
   getUserPosts,
-  isFollowing,
+  updateProfile,
 } from "@/actions/profile.action";
-import { notFound } from "next/navigation";
-import ProfilePageClient from "./ProfilePageClient";
-import { getCurrentUserId } from "@/actions/user.action";
-import Sidebar from "@/components/Sidebar";
+import { toggleFollow } from "@/actions/user.action";
+import { Avatar, AvatarImage } from "@/components/ui/avatar";
+import { Button } from "@/components/ui/button";
+import { Card, CardContent } from "@/components/ui/card";
+import {
+  Dialog,
+  DialogClose,
+  DialogContent,
+  DialogHeader,
+  DialogTitle,
+} from "@/components/ui/dialog";
+import { Input } from "@/components/ui/input";
+import { Label } from "@/components/ui/label";
+import { Separator } from "@/components/ui/separator";
+import { Tabs, TabsContent, TabsList, TabsTrigger } from "@/components/ui/tabs";
+import { Textarea } from "@/components/ui/textarea";
+import { SignInButton, useUser } from "@clerk/nextjs";
+import { format } from "date-fns";
+import {
+  CalendarIcon,
+  EditIcon,
+  FileTextIcon,
+  HeartIcon,
+  LinkIcon,
+  MapPinIcon,
+  ArrowLeftIcon,
+} from "lucide-react";
+import { useState } from "react";
+import toast from "react-hot-toast";
+import { useRouter } from "next/navigation";
+import ProfilePostFeed from "@/components/ProfilePostFeed";
 
-export async function generateMetadata({ params }: { params: { username: string } }) {
-  const user = await getProfileByUsername(params.username);
-  if (!user) return;
+type User = Awaited<ReturnType<typeof getProfileByUsername>>;
+type PostsData = Awaited<ReturnType<typeof getUserPosts>>;
 
-  return {
-    title: `${user.name ?? user.username}`,
-    description: user.bio || `Check out ${user.username}'s profile.`,
-  };
+interface ProfilePageProps {
+  user: NonNullable<User>;
+  posts: PostsData;
+  likedPosts: PostsData;
+  isFollowing: boolean;
+  currentUserId: string | null;
 }
 
-async function ProfilePageServer({ params }: { params: { username: string } }) {
-  const profileUser = await getProfileByUsername(params.username);
+export default function ProfilePage({
+  user,
+  posts,
+  likedPosts,
+  isFollowing: initialIsFollowing,
+  currentUserId,
+}: ProfilePageProps) {
+  const router = useRouter();
+  const { user: currentUser } = useUser();
 
-  if (!profileUser) notFound();
+  const [showEditDialog, setShowEditDialog] = useState(false);
+  const [isFollowing, setIsFollowing] = useState(initialIsFollowing);
+  const [isUpdatingFollow, setIsUpdatingFollow] = useState(false);
 
-  const [postsData, likedPostsData, isCurrentUserFollowing] = await Promise.all([
-    getUserPosts(profileUser.id),
-    getUserLikedPosts(profileUser.id),
-    isFollowing(profileUser.id),
-  ]);
-  const currentUserId = await getCurrentUserId();
+  const [editForm, setEditForm] = useState({
+    name: user.name || "",
+    bio: user.bio || "",
+    location: user.location || "",
+    website: user.website || "",
+  });
+
+  const handleEditSubmit = async () => {
+    const formData = new FormData();
+    Object.entries(editForm).forEach(([key, value]) =>
+      formData.append(key, value)
+    );
+
+    const result = await updateProfile(formData);
+    if (result.success) {
+      setShowEditDialog(false);
+      toast.success("Profile updated successfully");
+    }
+  };
+
+  const handleFollow = async () => {
+    if (!currentUser) return;
+    try {
+      setIsUpdatingFollow(true);
+      await toggleFollow(user.id);
+      setIsFollowing(!isFollowing);
+    } catch {
+      toast.error("Failed to update follow status");
+    } finally {
+      setIsUpdatingFollow(false);
+    }
+  };
+
+  const isOwnProfile =
+    currentUser?.username === user.username ||
+    currentUser?.emailAddresses[0].emailAddress.split("@")[0] === user.username;
+
+  const formattedDate = format(new Date(user.createdAt), "MMMM yyyy");
 
   return (
-    <div className="grid grid-cols-1 lg:grid-cols-12 gap-6">
-      <div className="hidden lg:block lg:col-span-3">
-        <Sidebar />
+    <div className="max-w-3xl mx-auto">
+      <div className="mb-4">
+        <Button variant="ghost" onClick={() => router.back()} className="flex items-center">
+          <ArrowLeftIcon className="size-4" />
+          Back
+        </Button>
       </div>
-      <div className="lg:col-span-6">
-        <ProfilePageClient
-          user={profileUser}
-          posts={postsData}
-          likedPosts={likedPostsData}
-          isFollowing={isCurrentUserFollowing}
-          currentUserId={currentUserId}
-        />
+
+      {/* Profile Card */}
+      <div className="grid grid-cols-1 gap-6">
+        <div className="w-full max-w-lg mx-auto">
+          <Card>
+            <CardContent className="pt-6">
+              <div className="flex flex-col items-center text-center">
+                <Avatar className="w-24 h-24">
+                  <AvatarImage src={user.image ?? "/avatar.png"} />
+                </Avatar>
+                <h1 className="mt-4 text-2xl font-bold">{user.name ?? user.username}</h1>
+                <p className="text-muted-foreground">@{user.username}</p>
+                <p className="mt-2 text-sm">{user.bio}</p>
+
+                <div className="w-full mt-6">
+                  <div className="flex justify-between mb-4">
+                    <div>
+                      <div className="font-semibold">{user._count.following}</div>
+                      <div className="text-sm text-muted-foreground">Following</div>
+                    </div>
+                    <Separator orientation="vertical" />
+                    <div>
+                      <div className="font-semibold">{user._count.followers}</div>
+                      <div className="text-sm text-muted-foreground">Followers</div>
+                    </div>
+                    <Separator orientation="vertical" />
+                    <div>
+                      <div className="font-semibold">{user._count.posts}</div>
+                      <div className="text-sm text-muted-foreground">Posts</div>
+                    </div>
+                  </div>
+                </div>
+
+                {!currentUser ? (
+                  <SignInButton mode="modal">
+                    <Button className="w-full mt-4">Follow</Button>
+                  </SignInButton>
+                ) : isOwnProfile ? (
+                  <Button className="w-full mt-4" onClick={() => setShowEditDialog(true)}>
+                    <EditIcon className="size-4 mr-2" />
+                    Edit Profile
+                  </Button>
+                ) : (
+                  <Button
+                    className="w-full mt-4"
+                    onClick={handleFollow}
+                    disabled={isUpdatingFollow}
+                    variant={isFollowing ? "outline" : "default"}
+                  >
+                    {isFollowing ? "Unfollow" : "Follow"}
+                  </Button>
+                )}
+
+                <div className="w-full mt-6 space-y-2 text-sm">
+                  {user.location && (
+                    <div className="flex items-center text-muted-foreground">
+                      <MapPinIcon className="size-4 mr-2" />
+                      {user.location}
+                    </div>
+                  )}
+                  {user.website && (
+                    <div className="flex items-center text-muted-foreground">
+                      <LinkIcon className="size-4 mr-2" />
+                      <a
+                        href={
+                          user.website.startsWith("http") ? user.website : `https://${user.website}`
+                        }
+                        className="hover:underline"
+                        target="_blank"
+                        rel="noopener noreferrer"
+                      >
+                        {user.website}
+                      </a>
+                    </div>
+                  )}
+                  <div className="flex items-center text-muted-foreground">
+                    <CalendarIcon className="size-4 mr-2" />
+                    Joined {formattedDate}
+                  </div>
+                </div>
+              </div>
+            </CardContent>
+          </Card>
+        </div>
+
+        {/* Tabs */}
+        <Tabs defaultValue="posts" className="w-full">
+          <TabsList className="w-full justify-start border-b rounded-none h-auto p-0 bg-transparent">
+            <TabsTrigger
+              value="posts"
+              className="flex items-center gap-2 rounded-none data-[state=active]:border-b-2 data-[state=active]:border-primary
+                px-6 font-semibold"
+            >
+              <FileTextIcon className="size-4" />
+              Posts
+            </TabsTrigger>
+            <TabsTrigger
+              value="likes"
+              className="flex items-center gap-2 rounded-none data-[state=active]:border-b-2 data-[state=active]:border-primary
+                px-6 font-semibold"
+            >
+              <HeartIcon className="size-4" />
+              Likes
+            </TabsTrigger>
+          </TabsList>
+
+          <TabsContent value="posts" className="mt-6">
+            {posts.posts.length > 0 ? (
+              <ProfilePostFeed
+                initialPosts={posts.posts}
+                initialCursor={posts.nextCursor}
+                userId={user.id}
+                currentUserId={currentUserId}
+                type="posts"
+              />
+            ) : (
+              <div className="text-center py-8 text-muted-foreground">No posts yet</div>
+            )}
+          </TabsContent>
+
+          <TabsContent value="likes" className="mt-6">
+            {likedPosts.posts.length > 0 ? (
+              <ProfilePostFeed
+                initialPosts={likedPosts.posts}
+                initialCursor={likedPosts.nextCursor}
+                userId={user.id}
+                currentUserId={currentUserId}
+                type="likes"
+              />
+            ) : (
+              <div className="text-center py-8 text-muted-foreground">No liked posts to show</div>
+            )}
+          </TabsContent>
+        </Tabs>
+
+        {/* Edit Profile Dialog */}
+        <Dialog open={showEditDialog} onOpenChange={setShowEditDialog}>
+          <DialogContent className="sm:max-w-[500px]">
+            <DialogHeader>
+              <DialogTitle>Edit Profile</DialogTitle>
+            </DialogHeader>
+            <div className="space-y-4 py-4">
+              <div className="space-y-2">
+                <Label>Name</Label>
+                <Input
+                  value={editForm.name}
+                  onChange={(e) => setEditForm({ ...editForm, name: e.target.value })}
+                  placeholder="Your name"
+                />
+              </div>
+              <div className="space-y-2">
+                <Label>Bio</Label>
+                <Textarea
+                  value={editForm.bio}
+                  onChange={(e) => setEditForm({ ...editForm, bio: e.target.value })}
+                  className="min-h-[100px]"
+                  placeholder="Tell us about yourself"
+                />
+              </div>
+              <div className="space-y-2">
+                <Label>Location</Label>
+                <Input
+                  value={editForm.location}
+                  onChange={(e) => setEditForm({ ...editForm, location: e.target.value })}
+                  placeholder="Where are you based?"
+                />
+              </div>
+              <div className="space-y-2">
+                <Label>Website</Label>
+                <Input
+                  value={editForm.website}
+                  onChange={(e) => setEditForm({ ...editForm, website: e.target.value })}
+                  placeholder="Your personal website"
+                />
+              </div>
+            </div>
+            <div className="flex justify-end gap-3">
+              <DialogClose asChild>
+                <Button variant="outline">Cancel</Button>
+              </DialogClose>
+              <Button onClick={handleEditSubmit}>Save Changes</Button>
+            </div>
+          </DialogContent>
+        </Dialog>
       </div>
     </div>
   );
 }
-export default ProfilePageServer;
